Fix the date formatting for answer results.
Forgot that I started this refactor and broke the build.  The idea is that while the UI may wish to display only certain parts of a date (month/year, month/day, hour/minute), the result is easier to normalize if it just puts in a value that is timeOnly or dateOnly with a standardized format.