Refactor the form and instruction step view controller to register table cells.
The previous design is based on ORKFormStepViewController and was architected to
dequeue a cell based on the indexPath rather than dequeuing cells using the newer
method that also includes the index path. By using the newer method, the OS will
track selection state and cell size, and this allows both registering table cells
in code and/or using a storyboard or nib to design custom versions of the view
controller.