Update the walk-and-balance and tremor navigation
Fix the “Restart Test” and “Review Instructions” buttons so that they will work for both Walk and Balance **and** Tremor. This fixes a bug where tapping on “Restart Test” in the tremor activity will reset the actions. This also allows the countdown to include a review instructions button and allows the buttons to be easily hidden if no action is associated with them.