Refactor image theme elements and UI actions to require `type` field.
JSON serialization on Android does not readily support drop-through failure cases for casting to different object types. Because of this, JSON files that are cross-platform need to use the “type” key in all cases where the object might need to support multiple cast objects.

Note: I wrapped throwing an error in a `#if DEBUG` for now so that release versions of the app won’t fail. Since our applications generally have unit tests that decode all the JSON files, this **should** catch these before release, but just in case, I didn’t want something to fail in production.