Deprecate “question” style JSON schema support.
Change the rules for the form step JSON schema to *only* support input fields as a list array. This is to
simplify decoding on Android and to produce more consistent standards for Decodable objects.

Additionally, allow the identifier for a input field to be inherited from its parent step if the JSON
schema does not explicitly define an identifier.