Change the `nextStepIdentifier` to immutable to avoid misuse
The steps in a task are intended to be immutable once the task has started so that the task can run without having to put in work-arounds for task navigation which may revisit the step either because of a navigation loop or because of “backward” navigation.

However, I am seeing this property being used in violation of that expectation. Since `RSDUIStepObject` is a class and not a struct, properties that are get/set will mutate the step rather than creating a new copy of the object. Unfortunately, because structs do not support inheritance, that makes this step as a **class** particularily useful (b/c it handles a lot of deserialization logic for you). While this isn’t an issue with the other properties, because the `nextStepIdentifier` will affect navigation, it should **only** be set during initialization.