Refactor framework to more closely map to Android implementation
This is a first step in refactoring the Research framework to more closely map to the organization of the
Android implementation.

Note: This does not include using ViewModel/Fragment or POJO implementations since those do not have an
iOS/watchOS/tvOS equivalent. It does split the "API", including protocols used during task presentation,
from the other sections of the code that include concrete implementations of the API protocols.