Refactor Research, ResearchUI, and MotorControl module to use the Sage Design System
While attempting to remain reverse-compatible to the previous color theme rules,
there are a lot of changes and deprication that was necessary to make the transition.
This is considered version 0 of the design system and attempts to keep to the rules
as estabilished for the previous apps rather than the updated rules.

Additionally, the color pallettes are likely to change again once the design team
has had an opportunity to see the results of the rules thus far.