Add custom views for a set of design system buttons that use template images
I made another hat.

This adds a button/view type for displaying a template image in a button or view. Currently, this is being used extensively by the design team for the Cognition Mobile Toolbox instruments.

I intentionally did not set up design color rules for this view b/c the current designs aren’t consistent enough to provide that.