Handle phone calls by stopping the recorder
Currently, if the participant receives a phone call while they are performing a task, the recorder will stop running. Additionally, the progress dial will pause/resume its countdown.

If the task is set up to stop on interruption, then the task view controller will fail the task and show the participant a message that the task has been stopped due to an interruption.

Note: Currently, the motion sensor recorder will not try to resume even if the phone call is ignored.