When a active step stops, stop the recorders
The designs for some of the active steps have changed to remain on the step until the participant
taps a “Next” button. Because of this, the recorders can continue running for a long time. Stop
the recorders when the step timer indicates that the step is done recording.

To allow for testing this, the default active step nib and UI/UX was changed to the active step used
by the MotorControl module. Currently, the default is not used in any modules that we support.