ME-110 Do not allow setting async recorder state to less than current
Because of the different threads used in starting and stopping async actions, and because there are various different timings for *when* a recorder is started/stopped, state handling can get confusing.

This change sets the motion recorder status to `requestingPermission` before checking the current permission state. Additionally, it only allows setting the requested permission (handled by the task controller) in a forward direction. (Will not set to “requestPermission” if already “starting”)