Set default presentation style to full screen modal
This is for an iOS 13 issue where a swipe down gesture can be used to dismiss a view controller. Setting to fulllscreen modal will remove that gesture recognizer. This is done during initialization so that it can be set by the presenting view controller to something else if the UI/UX calls for doing so.