Fix unit test for iOS 13
There appears to be a change with iOS 13 that this formatter no longer
conforms to the documentation for US English to read "1:30" versus "01:30".
In considering the "proper" behavior for this formatter, the formatting should
respect the zero padding everywhere if the `zeroFormattingBehavior` including
padding.

This updates the unit tests and updates the formatter for older iOS versions
to exhibit the same behavior for all iOS versions.