Add a app active listener to the motion authorization
Use the app notification for did become active to refresh the cached authorization whenever the app becomes active. This will handle the case where the participant has previously authorized access to the motion sensors, but has subsequently turned OFF that authorization.

Note: realistically, this is uncommon and will not cause a serious failure, but since testers will run this scenario, may as well fix it.