Implement Question/InputItem kotlin-supported serialization
Kotlin does not support the RSDFactory serialization that we are using in SageResearch so all the “type” class keys need to have a one-to-one mapping of the “type” key (which is required) and the associated class. The complex serialization strategies that we have employed in SageResearch will not work.

The most significant change is in the serialization of the form step and input fields. This commit addresses the serialization.