add “abandon assessment” action and process results
By request by MTB team, added a navigation action specific to
abandoning an assessment and allowing the task and/or navigator to
explicitly conform to a protocol to do post-processing on the results.

The idea behind the “abandonAssessment” action is to allow a button
to be set up to *upload* results and mark the task as finished without
actually finishing the task. This is different from how we tend to do UI/
UX where, for example, a participant may want to skip doing an
assessment and we use the navigator to skip to the end. What they
want is to allow a button push to kill the task *without* consulting the
navigator.

WRT the results post-processing. Again, they didn’t want to have the
navigator process the results when the task was completed or use a
custom task result with it’s own implementation of RSDArchivable.
Instead, they wanted to have a separate call in the taskViewModel
to process and mutate the task result outside of those provided
mechanisms.

In the interest of time and shortened schedules, I decided to just
go ahead and implement it.  ¯\_(ツ)_/¯